1.0.19 - League - checkAvailable league zwraca NULLa sprawdz TODO:  - dodac Test 10.2 min. 8:15.

+++ src/main/java/pl/com/k1313/g4g/domain/league/LeagueService.java
@@ -1,5 +1,6 @@
 package pl.com.k1313.g4g.domain.league;
 
+import com.mysql.cj.protocol.a.NativeConstants;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.com.k1313.g4g.domain.appuser.AppUser;
@@ -29,38 +30,43 @@ public class LeagueService {
 
         Club userClub = this.clubRepository.findByClubId(userClubId);
         League userLeague = checkAvailableLeague();
-        List<Club> leagueTeams = new ArrayList<>();
-        if (userLeague.getLeagueTeams() != null) {
-            leagueTeams = userLeague.getLeagueTeams();
-        }
-        int test = leagueTeams.size();
-//        tutaj trzeba sprawdzic czy zespoly ligii to boty i wyrzucic jeden jesli tak
-//            if (leagueTeams != null) {
-        for (Club c : leagueTeams
-        ) {
-            if (c.getAppUser() == null) {
-                this.clubRepository.delete(c);
-                break;
-            }
-        }
-        leagueTeams.add(userClub);
         userClub.setClubLeague(userLeague);
         this.clubRepository.save(userClub);
-        //tutaj musi cos userLeague sprawdzic stworzyc
-        this.leagueRepository.save(
-        while (userLeague.getLeagueTeams().size() < 8) {
-            this.clubRepository.save(this.clubService.botClubCreation());
-        }
-//            } else {
+        List<Club> leagueTeams=new ArrayList<>();
+        //to do zrpbienia w przyszlosci - co jesliw  availbele league jest juz uzytkownik
+//      if (!userLeague.getLeagueTeams().isEmpty()){leagueTeams=userLeague.getLeagueTeams();}
         leagueTeams.add(userClub);
+        this.leagueRepository.save(userLeague);
+//        tutaj trzeba sprawdzic czy zespoly ligi to boty i wyrzucic jeden jesli tak
+        if (leagueTeams.size() > 8) {
+            for (Club c : leagueTeams
+            ) {
+                if (c.getAppUser() == null) {
+                    this.clubRepository.delete(c);
+                    break;
+                }
+            }
+        }
+
+//        while (userLeague.getLeagueTeams().size() < 8) {
+//            this.clubRepository.save(this.clubService.botClubCreation());
+//        }
+////            } else {
         while (leagueTeams.size() < 8) {
             Club newBotClub = this.clubService.botClubCreation();
+;            newBotClub.setClubLeague(userLeague);
             leagueTeams.add(newBotClub);
             this.clubRepository.save(newBotClub);
         }
 //            }
 
         this.leagueRepository.save(userLeague);
+        //jak wydlubac zespoly z ligi??? tylko z leagueTeams??
+        System.out.println("----------------");
+        System.out.println("Liga: "+ userLeague.getId());
+        System.out.println("----------------");
+        System.out.println("zespoly: "+leagueTeams);
+        //zwraca NUlllaaaa!! ponizej
         return userLeague;
     }
 
@@ -69,7 +75,7 @@ public class LeagueService {
         List<League> leagues = this.leagueRepository.findAll();
         for (League l : leagues
         ) {
-            List<Club> leagueClubs = this.clubRepository.findByClubLeague(l);
+            List<Club> leagueClubs = this.clubRepository.findByClubLeague(l.getId());
             List<AppUser> leagueAppUsers = new ArrayList<>();
             for (Club c : leagueClubs
             ) {
@@ -79,25 +85,10 @@ public class LeagueService {
                         league = l;
                         //teraz wyciagnij nr ligi a jesli nie ma to nadaj pierwszy wolny
                         if (league.getLeagueNumber().isEmpty()) {
-                            List<League> leaguesWithNoNumbers =
-                                    this.leagueRepository.findAnyByLeagueNumberNotNull();
-                            if (!leaguesWithNoNumbers.isEmpty()) {
-                                int maxInt = 0;
-                                for (League leag : leaguesWithNoNumbers
-                                ) {
-                                    int leagueNumber = Integer.parseInt(leag.getLeagueNumber());
-                                    if (leagueNumber > maxInt) {
-                                        maxInt = leagueNumber;
-                                    }
-                                }
-                                String stringMaxInt=String.valueOf(maxInt+1);
-                                league.setLeagueNumber(stringMaxInt);
-                            }
-                        } else {
-
-                            dokonczyc tego elsa
+                            setLeagueNumber(league);
                         }
                     } else {
+                        league = l;
                         break;
                     }
                 }
@@ -106,5 +97,22 @@ public class LeagueService {
         this.leagueRepository.save(league);
         return league;
 
+    }
+
+    public String setLeagueNumber(League league) {
+        String stringMaxLeagueNr = "1";
+        List<League> leaguesWithNoNumbers =
+                this.leagueRepository.findAnyByLeagueNumberNotNull();
+        int maxInt = 0;
+        for (League leag : leaguesWithNoNumbers
+        ) {
+            int leagueNumberToSet = Integer.parseInt(leag.getLeagueNumber());
+            if (leagueNumberToSet > maxInt) {
+                maxInt = leagueNumberToSet;
+            }
+        }
+        stringMaxLeagueNr = String.valueOf(maxInt + 1);
+        league.setLeagueNumber(stringMaxLeagueNr);
+        return stringMaxLeagueNr;
     }
 }
